requirements/base.txt
Django==1.9.7
django-model-utils==2.5
Pillow==3.2.0
psycopg2==2.6.1
Unipath==1.1
